pyproject.toml
[tool.poetry]
name = "agentia"
version = "0.4.0"
description = "ChatGPT API wrapper, argumented with functions"
# documentation = "https://github.com/wenyuzhao/augmented-gpt"
authors = ["Wenyu Zhao <wenyu.zhao@anu.edu.au>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/wenyuzhao/augmented-gpt"
packages = [{ include = "agentia" }]
keywords = ["chatgpt", "gpt"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.0"
openai = "^1.21.2"
aiohttp = "^3.9.3"
jinja2 = "^3.1.3"
pillow = "^10.2.0"
pyright = "^1.1.361"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.2"
pytest = "^7.3.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
